split test into render & parse
